requirements.txt
# Core dependencies for MARBERT Turn Detector
livekit-agents>=1.3.9,<2.0.0
transformers>=4.30.0,<5.0.0
torch>=2.0.0,<3.0.0

# Automatically installed by above
huggingface-hub>=0.34.0
tokenizers>=0.20.0